# backend/src/config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "PwndocNG API Documentation",
    "description": "",
    "version": "1.0.0"
  },
  "host": "/",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/clients": {
      "get": {
        "tags": [
          "Client"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "Client"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "lastname": {
                  "example": "any"
                },
                "firstname": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "cell": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                },
                "company": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/clients/{id}": {
      "put": {
        "tags": [
          "Client"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "lastname": {
                  "example": "any"
                },
                "firstname": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "cell": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                },
                "company": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Client"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/companies": {
      "get": {
        "tags": [
          "Company"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "Company"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "shortName": {
                  "example": "any"
                },
                "logo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/companies/{id}": {
      "put": {
        "tags": [
          "Company"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "shortName": {
                  "example": "any"
                },
                "logo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Company"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
  }
}